requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==11.0.3

# AI and ML
openai>=1.12.0
transformers==4.35.2
torch==2.1.1
torchaudio==2.1.1
sentence-transformers==2.2.2
datasets==3.6.0

# Audio Processing
speechrecognition==3.10.0
SpeechRecognition==3.10.0

# Data Processing
pydantic>=2.7.0
pydantic-settings>=2.0.0
python-multipart==0.0.6
aiofiles==23.2.1
numpy==1.24.3
pandas==2.0.3
requests==2.31.0

# Document Processing
PyPDF2==3.0.1
python-docx==0.8.11
docling

# Optional packages (remove if causing conflicts)
# python-jose[cryptography]==3.3.0
# passlib[bcrypt]==1.7.4
